+++ packages/react-native-windows-init/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - react-native-windows-init
 
-This log was last generated on Wed, 20 Sep 2023 05:17:49 GMT and should not be manually modified.
+This log was last generated on Fri, 22 Sep 2023 05:13:33 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.3.42
+
+Fri, 22 Sep 2023 05:13:33 GMT
+
+### Patches
+
+- Bump @react-native-windows/telemetry to v0.0.0-canary.78
+- Bump @react-native-windows/cli to v0.0.0-canary.188
 
 ## 1.3.41